// src/app/backend/api/dashboard.service.ts
export interface ApiDashboardStatisticsGetRequestParams {
    dateFrom?: string | null;
    dateTo?: string | null;
}