// src/app/backend/model/promotionResponseDto.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimeSpan } from './timeSpan';
import { PromotionStatusEnum } from './promotionStatusEnum';
import { GeofenceResponseDto } from './geofenceResponseDto';
import { UserResponseDto } from './userResponseDto';
import { UserIdentityImageResponseDto } from './userIdentityImageResponseDto';


export interface PromotionResponseDto { 
    id?: number | null;
    name?: string | null;
    description?: string | null;
    startDate?: string;
    endDate?: string;
    startTime?: TimeSpan;
    endTime?: TimeSpan;
    weekDays?: Array<number> | null;
    geofence?: GeofenceResponseDto;
    geofenceId?: number | null;
    user?: UserResponseDto;
    userId?: number | null;
    createdOn?: string | null;
    design?: string | null;
    html?: string | null;
    image?: UserIdentityImageResponseDto;
    contact?: UserIdentityImageResponseDto;
    isEnabled?: boolean;
    isContinues?: boolean;
    subscriptionId?: number | null;
    status?: PromotionStatusEnum;
}