// src/app/backend/model/subscriptionFrequencyEnum.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type SubscriptionFrequencyEnum = 1 | 2 | 3 | 4;

export const SubscriptionFrequencyEnum = {
    Monthly: 1 as SubscriptionFrequencyEnum,
    HalfYearly: 2 as SubscriptionFrequencyEnum,
    Yearly: 3 as SubscriptionFrequencyEnum,
    Once: 4 as SubscriptionFrequencyEnum
};