// src/app/backend/model/subscriptionResponseDto.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SubscriptionFrequencyEnum } from './subscriptionFrequencyEnum';


export interface SubscriptionResponseDto { 
    id?: number;
    planName?: string | null;
    planId?: string | null;
    brandIdentityLimit?: number;
    brandIdentityUsed?: number;
    contactIdentityLimit?: number;
    contactIdentityUsed?: number;
    geofencesLimit?: number;
    geofencesUsed?: number;
    promotionsLimit?: number;
    promotionsUsed?: number;
    subscribedDate?: string;
    nextInvoiceDate?: string | null;
    nextInvoiceAmount?: number | null;
    frequency?: SubscriptionFrequencyEnum;
    cancelAt?: string | null;
}